pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "anthropic>=0.86.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.2",
    "google-api-python-client>=2.193.0",
    "google-auth>=2.49.1",
    "google-auth-oauthlib>=1.3.0",
    "httpx>=0.28.1",
    "mt-940>=4.30.0",
    "openpyxl>=3.1.5",
    "pdfplumber>=0.11.9",
    "pgvector>=0.3.6",
    "playwright>=1.58.0",
    "pydantic-settings>=2.13.1",
    "python-dateutil>=2.9.0.post0",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.48",
    "sse-starlette>=2.0.0",
    "uvicorn[standard]>=0.42.0",
    "weasyprint>=68.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
